Completely removing the AutoMapper, we use explicit conversion between types.

+++ Microting.DigitalOceanBase/Infrastructure/Data/Entities/DropletVersion.cs
@@ -5,15 +5,15 @@ namespace Microting.DigitalOceanBase.Infrastructure.Data.Entities
 {
     public class DropletVersion : BaseEntity
     {
-        public int DoUid { get; set; }
+        public long DoUid { get; set; }
         public int CustomerNo { get; set; }//// leave it empty
         public string PublicIpV4 { get; set; }
         public string PrivateIpV4 { get; set; }
         public string PublicIpV6 { get; set; }
         public string CurrentImageName { get; set; }
         public string RequestedImageName { get; set; }// should be filled during rebuild
-        public int CurrentImageId { get; set; }
-        public int RequestedImageId { get; set; }//// should be filled during rebuild
+        public long CurrentImageId { get; set; }
+        public long RequestedImageId { get; set; }//// should be filled during rebuild
         public string UserData { get; set; } // leave it empty
         public bool MonitoringEnabled { get; set; }
         public bool IpV6Enabled { get; set; }
